+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout-collection.component.spec.js
@@ -1,24 +1,24 @@
-import forEach from 'mocha-each';
+// import forEach from 'mocha-each';
 
-const Checkout = require('../../../test-utils/component-objects/f-checkout.component');
+// const Checkout = require('../../../test-utils/component-objects/f-checkout.component');
 
-const checkout = new Checkout();
+// const checkout = new Checkout();
 
-describe('f-checkout "collection" component tests', () => {
-    before(() => {
-        const checkoutData = {
-            type: 'collection',
-            isAuthenticated: true,
-            isValid: true
-        };
+// describe('f-checkout "collection" component tests', () => {
+//     before(() => {
+//         const checkoutData = {
+//             type: 'collection',
+//             isAuthenticated: true,
+//             isValid: true
+//         };
 
-        checkout.open(checkoutData);
-        checkout.waitForComponent();
-    });
+//         checkout.open(checkoutData);
+//         checkout.waitForComponent();
+//     });
 
-    forEach(['addressLine1', 'addressLine2', 'addressLocality', 'addressPostcode'])
-    .it.only('should check that address fields should not exist', field => {
-        // Assert
-        expect(checkout.doesFieldExist(field)).toBe(false);
-    });
-});
+//     forEach(['addressLine1', 'addressLine2', 'addressLocality', 'addressPostcode'])
+//     .it.only('should check that address fields should not exist', field => {
+//         // Assert
+//         expect(checkout.doesFieldExist(field)).toBe(false);
+//     });
+// });
